requirements.txt
certifi==2016.9.26
future==0.16.0
peewee==2.8.5
pkg-resources==0.0.0
python-telegram-bot==5.3.0
requests==2.12.4
urllib3==1.19.1
